Qt builds, but there seem to still be some define issues.
Mostly feature complete, but QT still causing pain. I belive that MacPorts
GNURadio is pinned to QT 4.4 which is already quite old. This is attempting
to use the most up-to-date int the 4.x series, which is 4.8.7 ("qt everywhere").

Indeed, Qt is all over the place. There are so many random patches necessary
to get this to build properly, this project is the poster-child for autotools.

I will curb my comments for now.

The pain that QT causes now is in the form of headers. I'm currently getting
errors of the form below, which exist in QT *headers* while trying to build
GNURadio.

/Applications/GNURadio.app/Contents/MacOS/usr/include/QtGui/qregion.h:57:1: error:
      expected unqualified-id
QT_MODULE(Gui)
^
/Applications/GNURadio.app/Contents/MacOS/usr/include/QtCore/qglobal.h:2899:5: note:
      expanded from macro 'QT_MODULE'
    typedef QtValidLicenseFor##x##Module Qt##x##Module;
    ^
I might continue to try and fix this, or just merge into master as-is, with
the hopes that wx-widgets are sufficient.