Use quoting in run-grc.sh
Hey, we're on Mac where typical users expect no issues with spaces in file or
directory names.

If an input argument in $@ has a space in it (such as a filesystem directory),
then $@ will expand that argument into multiple, thus preventing
gnuradio-companion from identifying the input.  But "$@" will preserve spaces
in arguments while still supporting $# != 1.  (This is different from "$*"
which will expand to only 1 argument regardless of $#, so "$*" and $* are not
correct either.)